Add Magnolia library with router, view and node modules
This commit introduces Magnolia, a new UI library for the web. The library includes core modules for routing, views and nodes that are fundamental components for a UI library. The new files also include a TypeScript config file, pipeline workflow script for Github actions, and a publish script for deploying the library.